Moved `MavenOperation` from types to enums
